fix: audit — 4 bug fixes across controllers and sales view
- PasswordController: trim() both sides + env() fallback to handle config cache edge case
- ReportController: (int) → (float) for inventory value (decimal qty bug for kg/litr)
- SaleController: stock error message now shows actual unit instead of hardcoded 'dona'
- sales/index: customer dropdown handles null company_name; cartCount badge 'dona'→'ta'

+++ app/Http/Controllers/ReportController.php
@@ -95,7 +95,7 @@ class ReportController extends Controller
         $inventoryValue = Product::whereNotNull('cost_price')
             ->where('cost_price', '>', 0)
             ->get()
-            ->sum(fn($p) => (float) $p->cost_price * (int) $p->quantity);
+            ->sum(fn($p) => (float) $p->cost_price * (float) $p->quantity);
 
         $inventoryCount = Product::count();
         $inventoryItems = Product::select('name', 'cost_price', 'quantity')
